finish log in page
