+++ frontend/src/App.css
@@ -14,6 +14,7 @@ html, body, #root {
     padding: 0;
     margin: 0;
     overflow: hidden; 
+    font-family: 'Helvetica', 'Arial', sans-serif;
     background-color: var(--background-color);
 }
