--- frontend/src/pages/Login.jsx
@@ -1,48 +1,0 @@
-import React, { useState } from "react";
-
-export default function LoginForm() {
-    const [username, setUsername] = useState("")
-    const [password, setPassword] = useState("")
-
-    const handleUsernameChange = (e) => {
-        setUsername(e.target.value)
-    }
-
-    const handlePasswordChange = (e) => {
-        setPassword(e.target.value)
-    }
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        console.log("Username:", username);
-        console.log("Password:", password);
-      };
-
-    return(
-        <form onSubmit={handleSubmit}>
-            <div>
-                <label htmlFor="username">Username:</label>
-                <input
-                    type="Text"
-                    id="username"
-                    value={username}
-                    onChange={handleUsernameChange}
-                    placeholder="Enter your username"
-                    required
-                />
-            </div>
-            <div>
-                <label htmlFor="password">Password:</label>
-                <input
-                    type="password"
-                    id="password"
-                    value={password}
-                    onChange={handlePasswordChange}
-                    placeholder="Enter your password"
-                    required
-                />
-            </div>
-            <button type="submit">Login</button>
-        </form>
-    );
-}
