--- frontend/src/pages/LoginForm.css
@@ -1,101 +1,0 @@
-/* General reset */
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    background: #f0f4fc; /* Background color */
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-
-.login-container {
-    display: flex;
-    width: 80%;
-    max-width: 1200px;
-    background: white;
-    border-radius: 15px;
-    overflow: hidden;
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);
-}
-
-.login-form {
-    flex: 1;
-    padding: 40px;
-}
-
-.login-form h2 {
-    font-size: 1.5rem;
-    color: #333;
-}
-
-.login-form p {
-    font-size: 1rem;
-    margin-bottom: 20px;
-    color: #666;
-}
-
-.input-group {
-    margin-bottom: 20px;
-}
-
-.input-group label {
-    display: block;
-    margin-bottom: 5px;
-    font-size: 0.9rem;
-    color: #555;
-}
-
-.input-group input {
-    width: 100%;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    font-size: 1rem;
-}
-
-.forgot-password {
-    margin-bottom: 20px;
-    text-align: right;
-}
-
-.forgot-password a {
-    text-decoration: none;
-    font-size: 0.9rem;
-    color: #007BFF;
-}
-
-button {
-    width: 100%;
-    padding: 10px;
-    font-size: 1rem;
-    background: #007BFF;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: background 0.3s ease;
-}
-
-button:hover {
-    background: #0056b3;
-}
-
-.login-illustration {
-    flex: 1;
-    background: linear-gradient(135deg, #b3d9ff, #007BFF);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 20px;
-}
-
-.login-illustration img {
-    max-width: 80%;
-    height: auto;
-}
